\name{CompLognormal-package}
\alias{CompLognormal-package}
\alias{CompLognormal}
\docType{package}
\title{
Computes functions for actuarial use
}
\description{Computes the probability density function, cumulative distribution function, quantile function, random numbers of any composite model based on the lognormal distribution}
\details{
\tabular{ll}{
Package: \tab ActuDistns\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2012-10-4\cr
License: \tab What license is it under?\cr
}
probability density function, cumulative distribution function, quantile function, random numbers
}
\author{Saralees Nadarajah

Maintainer: Saralees Nadarajah <Saralees.Nadarajah@manchester.ac.uk>
}
\references{
S. Nadarajah, S. A. A. Bakar, CompLognormal: An R Package for Composite Lognormal Distributions, submitted}
\keyword{ package }
